requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
altair==5.0.1
amqp==5.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
anyjson==0.3.3
appdirs==1.4.4
apsw==3.28.0.post1
apturl==0.5.2
arcaptcha==1.0.0
arcaptcha-python==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.1
asteroid-filterbanks==0.4.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==22.1.0
audioread==3.0.0
autopep8==2.0.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==4.2.0
bleach==5.0.1
blessed==1.20.0
blinker==1.4
Brlapi==0.7.0
Brotli==1.1.0
build==0.9.0
celery==5.4.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmaes==0.9.1
cmake==3.26.3
colorama==0.4.3
coloredlogs==15.0.1
colorlog==6.7.0
command-not-found==0.3
commonmark==0.9.1
ConfigArgParse==1.7
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
croniter==1.3.14
cryptography==2.8
css-parser==1.0.4
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
cycler==0.11.0
dateutils==0.6.12
dbus-python==1.2.16
dearpygui==1.9.1
debugpy==1.6.2
decorator==5.1.1
deepdiff==6.3.0
defer==1.0.6
defusedxml==0.7.1
discord.py==2.3.0
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2
django-celery==3.3.1
django-celery-results==2.5.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.19
drf-yasg==1.21.7
duplicity==0.8.12.0
einops==0.6.1
entrypoints==0.3
executing==0.8.3
fastapi==0.88.0
fastcluster==1.2.6
fasteners==0.14.1
fastjsonschema==2.16.1
feedparser==5.2.1
ffmpy==0.3.1
filelock==3.7.1
filterpy==1.4.5
fire==0.2.1
Flask==2.1.3
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
flatbuffers==23.5.9
fonttools==4.33.3
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
gevent==23.9.1
geventhttpclient==2.0.11
gradio==3.40.1
gradio_client==0.4.0
greenlet==2.0.2
h11==0.14.0
h5py==3.8.0
hawkauthlib==2.0.0
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.0.1
httpcore==0.16.3
httplib2==0.14.0
httpx==0.24.1
huggingface-hub==0.14.1
humanfriendly==10.0
HyperPyYAML==1.2.0
idna==2.8
ifaddr==0.1.6
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflection==0.5.1
inquirer==3.1.3
intervaltree==3.1.0
ipykernel==6.15.1
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.0.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.15.0
julius==0.2.7
jupyter-client==7.3.4
jupyter_core==5.7.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kaldi-io==0.9.5
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.4.2
kombu==5.3.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy_loader==0.2
librosa==0.10.0.post2
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.8.0
linkify-it-py==2.0.2
lit==16.0.5
llvmlite==0.40.0
lockfile==0.12.2
locust==2.17.0
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mechanize==0.4.5
mistune==2.0.4
mohawk==1.1.0
monotonic==1.5
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
nbclient==0.6.7
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.5
netifaces==0.10.4
networkx==3.1
nose==1.3.7
notebook==6.4.12
numba==0.57.0
numexpr==2.8.4
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.14.1
openapi-codec==1.3.2
optuna==3.1.1
ordered-set==4.1.0
orjson==3.9.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pep517==0.13.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2022.7.4
pipreqs==0.5.0
pkginfo==1.8.3
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pooch==1.6.0
primePy==1.3
prometheus-client==0.14.1
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio @ https://github.com/pyannote/pyannote-audio/archive/refs/heads/develop.zip
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pycairo==1.16.2
pychm==0.8.6
pycodestyle==2.10.0
pycparser==2.21
pycups==1.9.73
pydantic==1.10.7
pydub==0.25.1
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.8
python-dotenv==1.0.0
python-editor==1.0.4
python-magic==0.4.16
python-multipart==0.0.6
python-telegram-bot==20.4
pytorch-lightning==2.0.2
pytorch-metric-learning==2.1.1
pytz==2022.1
pyxdg==0.26
PyYAML==6.0
pyzmq==23.2.0
readchar==4.0.5
readme-renderer==37.3
redis==5.0.6
regex==2019.8.19
reportlab==3.5.34
repoze.lru==0.7
requests==2.28.1
requests-hawk==1.1.1
requests-toolbelt==0.10.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==12.6.0
roundrobin==0.0.4
Routes==2.4.1
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3cmd==2.0.2
scikit-learn==1.2.2
scipy==1.8.1
seaborn==0.11.2
SecretStorage==2.3.1
semantic-version==2.10.0
semver==3.0.0
Send2Trash==1.8.0
sentencepiece==0.1.99
shellingham==1.5.0.post1
simple-diarizer==0.0.13
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sklearn==0.0.post5
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==1.9.5
soxr==0.3.5
speechbrain==0.5.14
SQLAlchemy==2.0.13
sqlparse==0.4.4
stack-data==0.3.0
starlette==0.22.0
starsessions==1.3.0
stem==1.7.1
streamz==0.6.4
sympy==1.11.1
systemd-python==234
tabulate==0.9.0
tensorboardX==2.6
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.2
torchmetrics==0.11.4
tornado==6.2
tqdm==4.65.0
traitlets==5.3.0
traxtor==3.12
triton==2.0.0
twine==4.0.1
typer==0.9.0
typing_extensions==4.4.0
tzdata==2024.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.2
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.12
usb-creator==0.3.7
uvicorn==0.22.0
-e git+https://github.com/BUTSpeechFIT/VBx.git@7dfa3fab81c36b144cde9647fc80d96c0261a772#egg=VBx
vine==5.1.0
virtualenv==20.15.1
virtualenv-clone==0.5.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.5
webrtcvad==2.0.10
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.1.2
widgetsnbextension==4.0.2
xkit==0.0.0
yarg==0.1.9
yarl==1.9.2
zeroconf==0.24.4
zict==3.0.0
zipp==3.8.1
zope.event==5.0
zope.interface==6.1
